+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 ### Added
 - Format for `Set[A]`
+- Format for `Float`
 
 ## 0.1.3 - 2018-04-20
 ### Added